backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
PyYAML==6.0.2
jinja2==3.1.4
sqlmodel==0.0.21
psycopg2-binary==2.9.9
python-multipart==0.0.9

openai>=1.99.0
